Move the link highlight code to the engine,
so it could be shared, as decided at https://github.com/ckeditor/ckeditor5/pull/7356#issuecomment-637443530.
